modularize config
