+++ lua/user/plugins/vim_sleuth.lua
@@ -1,0 +1,4 @@
+-- Detect tabstop and shiftwidth automatically
+return {
+  { 'tpope/vim-sleuth' },
+}
